dfsaddsfasdasd

+++ src/main/java/org/example/springbootstripe/services/RolService.java
@@ -2,10 +2,12 @@ package org.example.springbootstripe.services;
 
 import org.example.springbootstripe.model.Rol;
 import org.example.springbootstripe.repository.RolRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RolService {
@@ -13,28 +15,22 @@ public class RolService {
     @Autowired
     private RolRepository rolRepository;
 
+    public Rol saveRol(Rol rol) {
+        return rolRepository.save(rol);
+    }
+
     public List<Rol> getAllRols() {
         return rolRepository.findAll();
     }
 
     public Rol getRolById(Long id) {
-        return rolRepository.findById(id).orElse(null);
-    }
-
-    public Rol saveRol(Rol usuari) {
-        return rolRepository.save(usuari);
-    }
-
-    public Rol updateRol(Long id, Rol usuari) {
-        Rol rolToUpdate = rolRepository.findById(id).orElse(null);
-        if (rolToUpdate != null) {
-            rolToUpdate.setNom(usuari.getNom());
-            return rolRepository.save(rolToUpdate);
-        }
-        return null;
+        Optional<Rol> rol = rolRepository.findById(id);
+        return rol.orElse(null);
     }
 
     public void deleteRol(Long id) {
-        rolRepository.deleteById(id);
+        if (rolRepository.existsById(id)) {
+            rolRepository.deleteById(id);
+        }
     }
 }
