+++ src/main/java/org/example/springbootstripe/services/UsuariService.java
@@ -2,10 +2,12 @@ package org.example.springbootstripe.services;
 
 import org.example.springbootstripe.model.Usuari;
 import org.example.springbootstripe.repository.UsuariRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UsuariService {
@@ -13,33 +15,30 @@ public class UsuariService {
     @Autowired
     private UsuariRepository usuariRepository;
 
+    public Usuari saveUsuari(Usuari usuari) {
+        return usuariRepository.save(usuari);
+    }
+
     public List<Usuari> getAllUsuaris() {
         return usuariRepository.findAll();
     }
 
     public Usuari getUsuariById(Long id) {
-        return usuariRepository.findById(id).orElse(null);
+        Optional<Usuari> usuari = usuariRepository.findById(id);
+        return usuari.orElse(null);
     }
 
-    public Usuari saveUsuari(Usuari usuari) {
-        return usuariRepository.save(usuari);
-    }
-
-    public Usuari updateUsuari(Long id, Usuari usuari) {
-        Usuari usuariToUpdate = usuariRepository.findById(id).orElse(null);
-        if (usuariToUpdate != null) {
-            usuariToUpdate.setNom(usuari.getNom());
-            usuariToUpdate.setCognoms(usuari.getCognoms());
-            usuariToUpdate.setNomUsuari(usuari.getNomUsuari());
-            usuariToUpdate.setEmail(usuari.getEmail());
-            usuariToUpdate.setContrasenya(usuari.getContrasenya());
-            usuariToUpdate.setIdRol(usuari.getIdRol());
-            return usuariRepository.save(usuariToUpdate);
+    public Usuari updateUsuari(Long id, Usuari updatedUsuari) {
+        if (usuariRepository.existsById(id)) {
+            updatedUsuari.setId(id);
+            return usuariRepository.save(updatedUsuari);
         }
         return null;
     }
 
     public void deleteUsuari(Long id) {
-        usuariRepository.deleteById(id);
+        if (usuariRepository.existsById(id)) {
+            usuariRepository.deleteById(id);
+        }
     }
 }
